tweaking syntax pre-release
