--- .github/workflows/ghDeploy2
@@ -1,34 +1,0 @@
-name: Deploy to GitHub Pages
-
-permissions:
-  contents: write
-
-on:
-  workflow_dispatch:
-  release:
-    types: [created]
-
-jobs:
-  build-and-deploy:
-    runs-on: [ubuntu-latest]
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: "20"
-
-      - name: Install Dependencies
-        run: npm ci
-
-      - name: Build
-        run: npm run build
-
-      - name: Deploy
-        uses: JamesIves/github-pages-deploy-action@4.1.5
-        with:
-          branch: gh-pages
-          folder: dist
